feat: exception handlers

+++ src/main/java/com/github_ol1veirx/smartcommerce/Services/ProductService.java
@@ -3,11 +3,16 @@ package com.github_ol1veirx.smartcommerce.Services;
 import com.github_ol1veirx.smartcommerce.DTO.ProductDTO;
 import com.github_ol1veirx.smartcommerce.Entities.Product;
 import com.github_ol1veirx.smartcommerce.Repository.ProductRepository;
+import com.github_ol1veirx.smartcommerce.Services.Exceptions.DatabaseException;
 import com.github_ol1veirx.smartcommerce.Services.Exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.lang.module.ResolutionException;
@@ -42,15 +47,32 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO) {
-        Product entity = productRepository.getReferenceById(id);
-        copyDtoToEntity(productDTO, entity);
-        entity = productRepository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDTO, entity);
+            entity = productRepository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch(EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Product Not Found");
+        }
+
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if(!productRepository.existsById(id)){
+            throw new ResourceNotFoundException("Product Not Found");
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (EmptyResultDataAccessException e) {
+            throw new ResourceNotFoundException("Product Not Found");
+        }
+        catch(DataIntegrityViolationException e) {
+            throw new DatabaseException("Integrity Violation");
+        }
     }
 
     private ProductDTO copyDtoToEntity(ProductDTO productDTO, Product entity) {
